requirements.txt
Flask
requests
pymysql
redis
pyjwt